refactor(inventory): 재주문 제안을 ReorderSuggestionReader 로 분리 (#724 항목 5-b)
읽기 전용·협력자 0·raw SQL 하나라 가장 깨끗하게 떨어진다. 본문 변경 0줄,
이름만 getReorderSuggestions → getSuggestions. 라우트·응답 타입 그대로.

#743(재고 보충 제안 재설계)의 작업 대상이 독립 파일이 된다. 지금 산식이 임시라는
사실(안전재고 리터럴 10, InboundPipelineReader 미사용)을 파일 docstring 에 적었다.

PurchaseOrderService: 733 → 678줄.
