+++ apps/core/src/modules/inventory/procurement/controllers/purchase-order.controller.ts
@@ -16,6 +16,7 @@ import { RequireScopes, ScopeGuard, User } from '@app/authorization';
 import { INVENTORY_SCOPE } from '../../../../platform/auth/inventory-scopes';
 import { PurchaseOrderService } from '../services/purchase-order.service';
 import { PurchaseOrderCartService } from '../services/purchase-order-cart.service';
+import { ReorderSuggestionReader } from '../services/reorder-suggestion.reader';
 import {
   CreatePurchaseOrderDto,
   UpdatePurchaseOrderStatusDto,
@@ -45,6 +46,7 @@ export class PurchaseOrderController {
   constructor(
     private readonly purchaseOrderService: PurchaseOrderService,
     private readonly cartService: PurchaseOrderCartService,
+    private readonly reorderReader: ReorderSuggestionReader,
   ) {}
 
   // ========== 발주 관리 ==========
@@ -196,7 +198,7 @@ export class PurchaseOrderController {
   })
   @ApiResponse({ status: 403, description: '재고 마스터데이터 관리 권한이 없습니다.' })
   async getReorderSuggestions(@Query('warehouseId') warehouseId?: string): Promise<StockReorderSuggestion[]> {
-    return this.purchaseOrderService.getReorderSuggestions(warehouseId);
+    return this.reorderReader.getSuggestions(warehouseId);
   }
 
   // ========== 발주 상세 조회 및 관리 (동적 라우트) ==========
