+++ apps/core/src/modules/inventory/procurement/procurement.module.ts
@@ -5,6 +5,7 @@ import { InboundModule } from '../inbound/inbound.module';
 import { PurchaseOrderController } from './controllers/purchase-order.controller';
 import { PurchaseOrderService } from './services/purchase-order.service';
 import { PurchaseOrderCartService } from './services/purchase-order-cart.service';
+import { ReorderSuggestionReader } from './services/reorder-suggestion.reader';
 
 /**
  * 조달(발주) 모듈. 경계는 ADR-0032 가 소유한다.
@@ -24,7 +25,7 @@ import { PurchaseOrderCartService } from './services/purchase-order-cart.service
 @Module({
   imports: [CoreInventoryModule, SharedModule, InboundModule],
   controllers: [PurchaseOrderController],
-  providers: [PurchaseOrderService, PurchaseOrderCartService],
-  exports: [PurchaseOrderService, PurchaseOrderCartService],
+  providers: [PurchaseOrderService, PurchaseOrderCartService, ReorderSuggestionReader],
+  exports: [PurchaseOrderService, PurchaseOrderCartService, ReorderSuggestionReader],
 })
 export class ProcurementModule {}
